# Enlarge the number of instructions in a basic block