Our experimental results show that the greater the number of instructions in a basic block, the greater the impact TCO has. We also discovered that if we use the branch instruction as the end of a basic block, there are only a few instructions in a basic block. To enlarge the number of instructions in a basic block, we only use a jump or call instruction as the end of the basic block, rather than a branch instruction.

CoreMark results:

| Model        | Compiler | TCO     |   Enlarge basic block   | Speedup |
|--------------+----------+---------+-------------------------+---------|
| Core i7-8700 | clang-15 | 971.951 |       1035.826899       |  +6.6%  |
|--------------+----------+---------+-------------------------+---------|
| Core i7-8700 | gcc-12   | 963.336 |       1123.895132       | +16.6%  |
|--------------+----------+---------+-------------------------+---------|
| eMAG 8180    | clang-15 | 335.396 |       383.819427        | +14.4%  |
|--------------+----------+---------+-------------------------+---------|
| eMAG 8180    | gcc-12   | 332.561 |       374.303071        | +12.5%  |