@qwe661234 2023-01-07
### CoreMark Result
| Model        | Compiler | TCO     | Enlarge Basic Block    | Speedup |
|--------------|----------|---------|---------|---------|
| Core i7-8700 | clang-15 | 971.951 | 1035.826899 | +6.6%  |
| Core i7-8700 | gcc-12   | 963.336 | 1123.895132 |  +16.6%  |
| eMAG 8180    | clang-15 | 335.396 | 383.819427 | +14.4%  |
| eMAG 8180    | gcc-12   | 332.561 | 374.303071 | +12.5%  |
### Compare the number of instructions in a basic block 
#### Model: Core i7-8700, Compiler: clang-15
* Before
```
Instruction |        Count        |       Invoked Times  
          1 |       257 [ 14.91 %]|        67811665 [  0.00 %]
          2 |       326 [ 18.91 %]|       128953458 [  0.00 %]
          3 |       287 [ 16.65 %]|       113370334 [  0.00 %]
          4 |       150 [  8.70 %]|        28792048 [  0.00 %]
          5 |       178 [ 10.32 %]|        87276451 [  0.00 %]
          6 |       102 [  5.92 %]|        42771905 [  0.00 %]
          7 |        74 [  4.29 %]|  94368206793558 [100.00 %]
          .
          .
          .
```
* After
```
Instruction |        Count        |       Invoked Times 
          0 |         0 [  0.00 %]|               0 [  0.00 %]
          1 |        43 [  3.48 %]|          221728 [  0.02 %]
          2 |        76 [  6.15 %]|        46721545 [  4.35 %]
          3 |       109 [  8.83 %]|         7935500 [  0.74 %]
          4 |        78 [  6.32 %]|        29186819 [  2.72 %]
          5 |        79 [  6.40 %]|        24730008 [  2.30 %]
          6 |        66 [  5.34 %]|        98268483 [  9.15 %]
          7 |        54 [  4.37 %]|        93258283 [  8.68 %]
          8 |        36 [  2.91 %]|        15350793 [  1.43 %]
          9 |        42 [  3.40 %]|         1424689 [  0.13 %]
         10 |        32 [  2.59 %]|         9619665 [  0.90 %]
         11 |        41 [  3.32 %]|         1017886 [  0.09 %]
         12 |        32 [  2.59 %]|        63336493 [  5.90 %]
         13 |        36 [  2.91 %]|        27690567 [  2.58 %]
         14 |        39 [  3.16 %]|        87729161 [  8.17 %]
         15 |        27 [  2.19 %]|        16175213 [  1.51 %]
         16 |        32 [  2.59 %]|         2152239 [  0.20 %]
         17 |        23 [  1.86 %]|        41594781 [  3.87 %]
         18 |        31 [  2.51 %]|        23869232 [  2.22 %]
         19 |        25 [  2.02 %]|        91090239 [  8.48 %]
         20 |        21 [  1.70 %]|         1145062 [  0.11 %]
         21 |        23 [  1.86 %]|          331961 [  0.03 %]
         22 |        13 [  1.05 %]|        10018980 [  0.93 %]
         23 |        13 [  1.05 %]|        10111048 [  0.94 %]
         24 |        12 [  0.97 %]|            2596 [  0.00 %]
         25 |        12 [  0.97 %]|        19443780 [  1.81 %]
         26 |         5 [  0.40 %]|              17 [  0.00 %]
         27 |        16 [  1.30 %]|        28597552 [  2.66 %]
         28 |        12 [  0.97 %]|        28894686 [  2.69 %]
         29 |         6 [  0.49 %]|         2834075 [  0.26 %]
         30 |        11 [  0.89 %]|          124481 [  0.01 %]
         31 |         3 [  0.24 %]|              10 [  0.00 %]
         32 |        13 [  1.05 %]|           49620 [  0.00 %]
         33 |         7 [  0.57 %]|          935059 [  0.09 %]
         34 |         4 [  0.32 %]|         7939518 [  0.74 %]
         35 |         5 [  0.40 %]|         4060710 [  0.38 %]
         36 |         6 [  0.49 %]|         3733205 [  0.35 %]
         37 |         2 [  0.16 %]|              12 [  0.00 %]
         38 |         6 [  0.49 %]|              75 [  0.00 %]
         39 |         4 [  0.32 %]|           93365 [  0.01 %]
         40 |         3 [  0.24 %]|            4242 [  0.00 %]
         41 |         8 [  0.65 %]|         6495489 [  0.60 %]
         42 |        15 [  1.21 %]|             494 [  0.00 %]
         43 |         6 [  0.49 %]|       182707629 [ 17.01 %]
         .
         .
         .
        
```