@jserv 2023-01-07
Move the experiments into #97 since the above is indeed a hack, violating the definition of basic block.
Later, we will proceed #104 and allow more optimizations.